+++ scripts/iso-manager.sh
@@ -13,18 +13,28 @@ SALTENV=""
 _usage() {
     echo "iso-manager.sh [options]"
     echo "Options:"
-    echo "-a/--archive <archive path>:     Path to archive folder or ISO"
-    echo "-l/--log-file <logfile_path>:    Path to log file"
-    echo "-v/--verbose:                    Run in verbose mode"
-    echo "-d/--dry-run:                    Run actions in dry run mode"
+    echo "-a/--add-archive <archive path>:     Path to an archive folder or ISO to add"
+    echo "    [DEPRECATED] You may also use --archive as an equivalent to the above."
+    echo "                 This option name will be removed in MetalK8s 124.0.0"
+    echo "-r/--rm-archive <archive path>:      Path to an archive folder or ISO to remove"
+    echo "-l/--log-file <logfile_path>:        Path to log file"
+    echo "-v/--verbose:                        Run in verbose mode"
+    echo "-d/--dry-run:                        Run actions in dry run mode"
 }
 
-ARCHIVES=()
+ARCHIVES_ADD=()
+ARCHIVES_RM=()
+ARCHIVE_DEPRECATED_OPTION=
 
 while (( "$#" )); do
   case "$1" in
-    -a|--archive)
-      ARCHIVES+=("$(readlink -f "$2")")
+    -a|--archive|--add-archive)
+      ARCHIVE_DEPRECATED_OPTION="true"
+      ARCHIVES_ADD+=("$(readlink -f "$2")")
+      shift 2
+      ;;
+    -r|--rm-archive)
+      ARCHIVES_RM+=("$(readlink -f "$2")")
       shift 2
       ;;
     -d|--dry-run)
@@ -47,6 +57,13 @@ while (( "$#" )); do
       ;;
   esac
 done
+
+if [ -z ${ARCHIVE_DEPRECATED_OPTION} ]; then
+  echo "Warning: Usage of the '--archive' option is deprecated, and will be removed" \
+       "in the next MetalK8s major version, 124.0.0. Please make sure to update your" \
+       "documentation and automation to use '--add-archive' instead (or the shorthand" \
+       "'-a' variant)." >&2
+fi
 
 TMPFILES=$(mktemp -d)
 
@@ -84,6 +101,12 @@ _add_archives() {
     done
 }
 
+_remove_archives() {
+    for archive; do
+        $SALT_CALL metalk8s.configure_archive "$archive" remove=True
+    done
+}
+
 _configure_archives() {
     # Mount archives
     echo "Mounting archives..."
@@ -115,6 +138,9 @@ _configure_archives() {
     echo "Making new versions available"
     $SALT_CALL state.sls metalk8s.archives saltenv="$SALTENV" test="$DRY_RUN" \
         --retcode-passthrough
+    # Unmount
+    $SALT_CALL state.sls metalk8s.archives.unmounted test="$DRY_RUN" \
+        --retcode-passthrough
 }
 
 _check_config() {
@@ -129,7 +155,10 @@ _set_env
 [ -z "$SALTENV" ] && die "saltenv not set"
 
 run "Check bootstrap configuration file" _check_config
-if (( ${#ARCHIVES[@]} )); then
-    run "Add archives" _add_archives "${ARCHIVES[@]}"
+if (( ${#ARCHIVES_ADD[@]} )); then
+    run "Add archives" _add_archives "${ARCHIVES_ADD[@]}"
+fi
+if (( ${#ARCHIVES_RM[@]} )); then
+    run "Remove archives" _remove_archives "${ARCHIVES_RM[@]}"
 fi
 run "Configure archives" _configure_archives